requirements.txt
cx_freeze
lapx
numpy
opencv-python
pillow
pygame
pygubu
python-dotenv
requests
scikit-image
ultralytics
yt-dlp